binary/v1.1_esm/requirements.txt
# ARG Binary Classification v1.1 - Requirements
# Python >= 3.8

torch>=2.0.0
transformers>=4.30.0
biopython>=1.81
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
tqdm>=4.65.0
